+++ server/jest.config.cjs
@@ -1,0 +1,5 @@
+module.exports = {
+    testEnvironment: "node",
+    testMatch: ["**/tests/**/*.test.js"],
+    collectCoverageFrom: ["**/*.js", "!**/node_modules/**"],
+  };
